Add presence-or validator
This validator checks for the presence of one or more values based on the attribute tied to the validator and the `dependentKeys` array passed into `options`.

Users can also pass an `exclusive: true` option that makes this an exclusive-or check. That is, it expects the presence of one and only one of the values being watched.