Fix for data validator with format
The date validator is not working with format specified.
It always says invalid value.

It is first validated without the format, returns false, since Date() cannot parse it and then will return an error message and never tries with the format.

This commit skips the Date() check if format is specified
and fixes the data validator test to use a format that is not parsed by
Date()